requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.2
python-dotenv==0.21.0
torch
torchvision
Pillow==9.5.0
Werkzeug==2.2.2
PyJWT==2.4.0
